refactor: add LikeButton
