+++ src/components/Posts/PostCardFooter.tsx
@@ -7,25 +7,30 @@ import { useTranslation } from "react-i18next";
 import { isLoggedInVar } from "../../apollo/cache";
 import {
   PostCardFooterFragment,
-  useDeleteLikeMutation,
-  useLikePostMutation,
   usePostCommentsLazyQuery,
 } from "../../apollo/gen";
-import { TypeNames } from "../../constants/common.constants";
-import { Blurple } from "../../styles/theme";
 import { inDevToast } from "../../utils/common.utils";
 import CommentForm from "../Comments/CommentForm";
 import CommentsList from "../Comments/CommentList";
 import CardFooterButton from "../Shared/CardFooterButton";
 import Flex from "../Shared/Flex";
 import { BASE_BADGE_STYLES } from "../Votes/VoteBadge";
+import LikeButton from "./LikeButton";
 
-const ICON_STYLES: SxProps = {
+export const ICON_STYLES: SxProps = {
   marginRight: "0.4ch",
 };
+
 const ROTATED_ICON_STYLES: SxProps = {
   ...ICON_STYLES,
   transform: "rotateY(180deg)",
+};
+
+const BADGE_STYLES: SxProps = {
+  ...BASE_BADGE_STYLES,
+  width: 22.5,
+  height: 22.5,
+  marginRight: 0.9,
 };
 
 interface Props {
@@ -34,50 +39,10 @@ interface Props {
 
 const PostCardFooter = ({ post: { id, likesCount, isLikedByMe } }: Props) => {
   const isLoggedIn = useReactiveVar(isLoggedInVar);
-
-  const [likePost, { loading: likePostLoading }] = useLikePostMutation();
-  const [unlikePost, { loading: unlikePostLoading }] = useDeleteLikeMutation();
-
   const [getPostComments, { data: postCommentsData }] =
     usePostCommentsLazyQuery();
 
   const { t } = useTranslation();
-
-  const isLoading = likePostLoading || unlikePostLoading;
-
-  const badgeStyles: SxProps = {
-    ...BASE_BADGE_STYLES,
-    width: 22.5,
-    height: 22.5,
-    marginRight: 0.9,
-  };
-  const likeButtonIconStyles: SxProps =
-    isLikedByMe && !isLoading
-      ? {
-          ...ICON_STYLES,
-          color: Blurple.Marina,
-        }
-      : ICON_STYLES;
-
-  const handleLikeButtonClick = async () => {
-    const variables = { likeData: { postId: id } };
-    if (isLikedByMe) {
-      unlikePost({
-        variables,
-        update(cache) {
-          cache.modify({
-            id: cache.identify({ __typename: TypeNames.Post, id }),
-            fields: {
-              isLikedByMe: () => false,
-              likesCount: (existingCount: number) => existingCount - 1,
-            },
-          });
-        },
-      });
-      return;
-    }
-    await likePost({ variables });
-  };
 
   const handleCommentButtonClick = async () =>
     await getPostComments({ variables: { id, isLoggedIn } });
@@ -87,7 +52,7 @@ const PostCardFooter = ({ post: { id, likesCount, isLikedByMe } }: Props) => {
       <Box paddingX="16px">
         {!!likesCount && (
           <Flex marginBottom={0.8}>
-            <Box sx={badgeStyles}>
+            <Box sx={BADGE_STYLES}>
               <LikeIcon
                 color="primary"
                 sx={{ fontSize: 13, marginTop: 0.65 }}
@@ -100,14 +65,7 @@ const PostCardFooter = ({ post: { id, likesCount, isLikedByMe } }: Props) => {
       </Box>
 
       <CardActions sx={{ justifyContent: "space-around" }}>
-        <CardFooterButton
-          sx={isLikedByMe ? { color: Blurple.Marina } : {}}
-          disabled={isLoading}
-          onClick={handleLikeButtonClick}
-        >
-          <LikeIcon sx={likeButtonIconStyles} />
-          {t("actions.like")}
-        </CardFooterButton>
+        <LikeButton postId={id} isLikedByMe={!!isLikedByMe} />
 
         <CardFooterButton onClick={handleCommentButtonClick}>
           <Comment sx={ROTATED_ICON_STYLES} />
